+++ pandas/core/categorical.py
@@ -21,7 +21,7 @@ from pandas.core.common import (
     _possibly_infer_to_datetimelike, get_dtype_kinds, is_list_like,
     is_sequence, is_null_slice, is_bool, _ensure_object, _ensure_int64,
     _coerce_indexer_dtype, take_1d)
-from pandas.core.dtypes import CategoricalDtype
+from pandas.types.api import CategoricalDtype
 from pandas.util.terminal import get_terminal_size
 from pandas.core.config import get_option
