+++ pandas/types/api.py
@@ -1,0 +1,2 @@
+from pandas.types.dtypes import (CategoricalDtype, CategoricalDtypeType,
+                                 DatetimeTZDtype, DatetimeTZDtypeType) # noqa
